+++ lib/connection/response.py
@@ -16,43 +16,29 @@
 #
 #  Author: Mauro Soria
 
+import httpx
+
 from lib.core.settings import (
-    DEFAULT_ENCODING, ITER_CHUNK_SIZE,
-    MAX_RESPONSE_SIZE, UNKNOWN,
+    DEFAULT_ENCODING,
+    ITER_CHUNK_SIZE,
+    MAX_RESPONSE_SIZE,
+    UNKNOWN,
 )
 from lib.parse.url import clean_path, parse_path
 from lib.utils.common import is_binary
 
 
-class Response:
+class BaseResponse:
     def __init__(self, response):
-        self.url = response.url
-        self.full_path = parse_path(response.url)
+        self.url = str(response.url)
+        self.full_path = parse_path(self.url)
         self.path = clean_path(self.full_path)
         self.status = response.status_code
         self.headers = response.headers
         self.redirect = self.headers.get("location") or ""
-        self.history = [res.url for res in response.history]
+        self.history = [str(res.url) for res in response.history]
         self.content = ""
         self.body = b""
-
-        for chunk in response.iter_content(chunk_size=ITER_CHUNK_SIZE):
-            self.body += chunk
-
-            if len(self.body) >= MAX_RESPONSE_SIZE or (
-                "content-length" in self.headers and is_binary(self.body)
-            ):
-                break
-
-        if not is_binary(self.body):
-            try:
-                self.content = self.body.decode(
-                    response.encoding or DEFAULT_ENCODING, errors="ignore"
-                )
-            except LookupError:
-                self.content = self.body.decode(
-                    DEFAULT_ENCODING, errors="ignore"
-                )
 
     @property
     def type(self):
@@ -77,3 +63,47 @@ class Response:
             other.body,
             other.redirect,
         )
+
+
+class Response(BaseResponse):
+    def __init__(self, response):
+        super().__init__(response)
+
+        for chunk in response.iter_content(chunk_size=ITER_CHUNK_SIZE):
+            self.body += chunk
+
+            if len(self.body) >= MAX_RESPONSE_SIZE or (
+                "content-length" in self.headers and is_binary(self.body)
+            ):
+                break
+
+        if not is_binary(self.body):
+            try:
+                self.content = self.body.decode(
+                    response.encoding or DEFAULT_ENCODING, errors="ignore"
+                )
+            except LookupError:
+                self.content = self.body.decode(DEFAULT_ENCODING, errors="ignore")
+
+
+class AsyncResponse(BaseResponse):
+    @classmethod
+    async def create(cls, response: httpx.Response) -> "AsyncResponse":
+        self = cls(response)
+        async for chunk in response.aiter_bytes(chunk_size=ITER_CHUNK_SIZE):
+            self.body += chunk
+
+            if len(self.body) >= MAX_RESPONSE_SIZE or (
+                "content-length" in self.headers and is_binary(self.body)
+            ):
+                break
+
+        if not is_binary(self.body):
+            try:
+                self.content = self.body.decode(
+                    response.encoding or DEFAULT_ENCODING, errors="ignore"
+                )
+            except LookupError:
+                self.content = self.body.decode(DEFAULT_ENCODING, errors="ignore")
+
+        return self
